#15: money representation — float64 rupees to int64 paise
Ships on its own, per the plan. All monetary fields move from float64/DECIMAL
to int64/BIGINT integer minor units (paise); the JSON wire format becomes an
integer count of paise. tax_rate stays a float percentage and quantity stays a
float count — neither is money.

- 000012_money_to_paise migration: DECIMAL -> BIGINT via ROUND(col*100),
  reversible; DEFAULT 0.00 columns handled
- models, DTOs, repositories switched to int64; SUM(...)::BIGINT in the ledger
- quotation calc uses integer math with explicit rounding; PDF formats paise/100
- dev_seed.sql money values converted to paise
- wave3_money_test pins the integer math (json.Number rejects rupee floats)

+++ backend/internal/bff/model.go
@@ -14,7 +14,7 @@ type ProjectCard struct {
 	ID          int       `json:"id"`           // Master Lead ID
 	ClientName  string    `json:"client_name"`
 	Status      string    `json:"status"`       // The current unified stage
-	Value       float64   `json:"value"`        // Aggregated from approved quote (if any)
+	Value       int64     `json:"value"`        // paise; aggregated from approved quote (if any)
 	LastUpdated time.Time `json:"last_updated"`
 }
 
@@ -53,7 +53,7 @@ type ActionItem struct {
 	Type        string    `json:"type"`         // "QUOTE_APPROVAL", "LEAD_ASSIGNMENT"
 	Title       string    `json:"title"`
 	RequestedBy string    `json:"requested_by"` // User ID or Name
-	Amount      float64   `json:"amount"`       // For expenses/quotes. 0 for leaves.
+	Amount      int64     `json:"amount"`       // paise; for expenses/quotes. 0 for leaves.
 	CreatedAt   time.Time `json:"created_at"`
 }
