+++ backend/internal/execution/model/execution_models.go
@@ -7,7 +7,7 @@ type Installer struct {
 	Name                 string    `json:"name" db:"name"`
 	Phone                string    `json:"phone" db:"phone"`
 	ExpertiseArea        *string   `json:"expertise_area" db:"expertise_area"`
-	StandardRate         *float64  `json:"standard_rate" db:"standard_rate"`
+	StandardRate         *int64    `json:"standard_rate" db:"standard_rate"` // paise
 	PreferredPaymentMode *string   `json:"preferred_payment_mode" db:"preferred_payment_mode"`
 	IsActive             bool      `json:"is_active" db:"is_active"`
 	CreatedAt            time.Time `json:"created_at" db:"created_at"`
@@ -18,13 +18,13 @@ type Installation struct {
 	OrderID                 int        `json:"order_id" db:"order_id"`
 	TechnicalManagerID      int        `json:"technical_manager_id" db:"technical_manager_id"`
 	InstallerID             *int       `json:"installer_id" db:"installer_id"`
-	AgreedInstallerPrice    *float64   `json:"agreed_installer_price" db:"agreed_installer_price"`
+	AgreedInstallerPrice    *int64     `json:"agreed_installer_price" db:"agreed_installer_price"` // paise
 	StartDate               *time.Time `json:"start_date" db:"start_date"`
 	EstimatedCompletionDate *time.Time `json:"estimated_completion_date" db:"estimated_completion_date"`
 	Status                  string     `json:"status" db:"status"`
 	InstallerJobStatus      string     `json:"installer_job_status" db:"installer_job_status"`
-	InstallerAdvanceAmount  float64    `json:"installer_advance_amount" db:"installer_advance_amount"`
-	InstallerFinalAmount    float64    `json:"installer_final_amount" db:"installer_final_amount"`
+	InstallerAdvanceAmount  int64      `json:"installer_advance_amount" db:"installer_advance_amount"` // paise
+	InstallerFinalAmount    int64      `json:"installer_final_amount" db:"installer_final_amount"`     // paise
 	ClientSignoffURL        *string    `json:"client_signoff_url" db:"client_signoff_url"`
 	ClientFeedback          *string    `json:"client_feedback" db:"client_feedback"`
 	CreatedAt               time.Time  `json:"created_at" db:"created_at"`
@@ -61,7 +61,7 @@ type InstallerPayment struct {
 	InstallationID       int       `json:"installation_id" db:"installation_id"`
 	InstallerID          int       `json:"installer_id" db:"installer_id"`
 	ProcessedBy          int       `json:"processed_by" db:"processed_by"`
-	Amount               float64   `json:"amount" db:"amount"`
+	Amount               int64     `json:"amount" db:"amount"` // paise
 	PaymentType          string    `json:"payment_type" db:"payment_type"`
 	PaymentMode          string    `json:"payment_mode" db:"payment_mode"`
 	TransactionReference *string   `json:"transaction_reference,omitempty" db:"transaction_reference"`
